- added level & exp to PlayerData class - added functions to freeze/reset animations for SpriteAnimator script - added speech bubble animation when mouse over an NPC - added sprites for Potions & Iris Fragments (the crystals/gems)

+++ Assets/Scripts/Game HUD/GameHUD.cs
@@ -33,12 +33,12 @@ public class GameHUD : MonoBehaviour {
 	
 	}
 
-    public void HPChanged(float newHP, float maxHP)
+    public void HPChanged(int newHP, int maxHP)
     {
         HPBar.localScale = new Vector3(newHP / maxHP, 1f, 1f);
     }
 
-    public void MPChanged(float newMP, float maxMP)
+    public void MPChanged(int newMP, int maxMP)
     {
         MPBar.localScale = new Vector3(newMP / maxMP, 1f, 1f);
     }
